+++ positions/src/modules/index.ts
@@ -1,2 +1,3 @@
 export {PositionToDirective} from "./positions/directives";
 export {PositionsModule} from "./positions/modules";
+export {TOOLTIP_OPTIONS, TooltipComponent, TooltipDirective, TooltipModule, TooltipOptions, TooltipRenderer} from './tooltip';