+++ positions/src/modules/tooltip/index.ts
@@ -1,0 +1,4 @@
+export * from './modules';
+export * from './directives';
+export * from './components';
+export * from './misc';