+++ positions/src/modules/tooltip/misc/tooltip.interface.ts
@@ -1,0 +1,72 @@
+import {TemplateRef, Type} from '@angular/core';
+
+/**
+ * Represents component that is used for rendering tooltip
+ */
+export interface TooltipRenderer<TData = any>
+{
+    //######################### properties #########################
+
+    /**
+     * Data that are rendered in tooltip
+     */
+    data: TData|null;
+
+    /**
+     * Template used for rendering tooltip
+     */
+    template: TemplateRef<TData>|null;
+
+    /**
+     * Indication whether are html tags allowed in tooltip text
+     */
+    allowHtml: boolean;
+
+    /**
+     * Css class that is applied to tooltip renderer component
+     */
+    cssClass: string|null;
+
+    //######################### methods #########################
+
+    /**
+     * Explicitly runs invalidation of content (change detection)
+     */
+    invalidateVisuals(): void;
+}
+
+/**
+ * Options used for tooltip directive
+ */
+export interface TooltipOptions
+{
+    /**
+     * Delay for displaying of tooltip on hover
+     */
+    delay?: number;
+
+    /**
+     * Indication whether is tooltip displayed at fixed position, or if it is displayed where cursor enters element
+     */
+    fixedPosition?: boolean;
+
+    /**
+     * Position where should be tooltip displayed at element
+     */
+    elementPositionAt?: Positions.PositionsCoordinates;
+
+    /**
+     * Position of tooltip where should placed at
+     */
+    tooltipPosition?: Positions.PositionsCoordinates;
+
+    /**
+     * Css class that is applied to tooltip renderer component
+     */
+    tooltipCssClass?: string|null;
+
+    /**
+     * Type of tooltip renderer that is used for rendering tooltip
+     */
+    tooltipRenderer?: Type<TooltipRenderer>;
+}