+++ positions/src/modules/tooltip/modules/tooltip.module.ts
@@ -1,0 +1,28 @@
+import {NgModule} from "@angular/core";
+import {CommonModule} from '@angular/common';
+
+import {TooltipComponent} from '../components';
+import {TooltipDirective} from '../directives';
+
+/**
+ * Module for rendering tooltips
+ */
+@NgModule(
+{
+    imports:
+    [
+        CommonModule
+    ],
+    declarations:
+    [
+        TooltipDirective,
+        TooltipComponent
+    ],
+    exports:
+    [
+        TooltipDirective
+    ]
+})
+export class TooltipModule
+{
+}